requirements.txt
beautifulsoup4==4.7.1
bsedata==0.2.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
dateutils==0.6.6
Flask==1.0.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
kafka-python==1.4.4
MarkupSafe==1.1.1
nsetools==1.0.8
python-constraint==1.4.0
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
six==1.12.0
soupsieve==1.8
urllib3==1.24.1
Werkzeug==0.14.1
